[2026-03-04] Some restructuring; dependency updates
Details:
- Moved $breakpoints from config to tokens
- Removed functions - mapToObject(), mapToObjectAsync() (now in utility-typescript instead)
- Removed type - RecursiveRecord (now in utility-typescript instead)
- Added config function - enforce-type-or-warn
- Updated with utility-typescript

+++ .scripts/classes/Compile.js
@@ -116,14 +116,16 @@ export class Compile extends CompileStage {
             1,
         );
 
-        this.console.verbose( 'tidying up compiled files...', 2 );
-        this.try(
-            this.fs.delete,
-            ( this.params.verbose ? 3 : 2 ),
-            [ [
-                'dist/css/template/@template.css',
-                'dist/css/template/@template.css.map'
-            ], ( this.params.verbose ? 3 : 2 ) ]
-        );
+        if ( this.params.packaging || this.params.releasing ) {
+            this.console.verbose( 'tidying up compiled files...', 2 );
+            this.try(
+                this.fs.delete,
+                ( this.params.verbose ? 3 : 2 ),
+                [ [
+                    'dist/scss/template/default-sample',
+                    'dist/scss/template/demos',
+                ], ( this.params.verbose ? 3 : 2 ) ]
+            );
+        }
     }
 }
